Completely refactor app routing scheme
